pyproject.toml
[project]
name = "application"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "logging>=0.4.9.6",
    "pathlib>=1.0.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
]

[project.optional-dependencies]
dev = ["ruff"]
test = ["pytest>=7.4.0", "pytest-asyncio>=0.22.0"]


[project.scripts]
runapp = "application.__main__:main"

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ruff>=0.12.7",
]
